Night Version
Massive restructuring of the entire project.
Ported the C# code to C++ so that the installer doesn't require the
dotnet framework to be present in order to run.

New licence: HAVE MERCY ON ME PUBLIC LICENCE.
Added the Windows licence to the proprietary bits and separated
them into their own special directory.

The C# bit isn't going to use evil msbuild anymore.
Currently in the process of creating more Makefiles to automate
the process. Gotta say, Rosyln C# compilers are pretty nice :D.

I'll try to get this finished by tomorrow and maybe publish a new
release. The C++ Makefile could use some cleaning. Well, the entire
project could use a bit of QA, not that I consider myself to be a
professional, by any means.